Split the library into multiple files
