+++ fastapi_injector/exceptions.py
@@ -1,0 +1,2 @@
+class InjectorNotAttached(ValueError):
+    """Raised when Injected is used without an injector instance attached to the app."""
